requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-MySQLdb==2.0.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python==8.4.0
mysqlclient==2.2.4
openai==1.35.4
pydantic==2.7.4
pydantic_core==2.18.4
PyJWT==2.8.0
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.4
twilio==9.2.1
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
